% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnFA.R
\name{MPLNFAClust}
\alias{MPLNFAClust}
\title{Clustering via Parsimonious Mixtures of MPLN Factor Analyzers Family}
\usage{
MPLNFAClust(
  dataset,
  membership = "none",
  gmin = 1,
  gmax = 3,
  pmin = 1,
  pmax = 2,
  initMethod = "kmeans",
  modelNames = "CCC",
  normalize = "Yes"
)
}
\arguments{
\item{dataset}{A dataset of class matrix and type integer such that
rows correspond to observations and columns correspond to variables.
The dataset have dimensions n x d, where n is the total number of
observations and d is the dimensionality. The program will cluster
n observations into gmin:gmax groups. If rowSums are zero,
these rows should be removed first.}

\item{membership}{A numeric vector of length nrow(dataset) containing the
cluster membership of each observation. If not available,
leave as "none".}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >= gmin, specifying the maximum number of
components to be considered in the clustering run.}

\item{pmin}{A positive integer specifying the minimum number of latent
factors to be considered in the clustering run.}

\item{pmax}{A positive integer, >= pmin, specifying the maximum number of
latent factors to be considered in the clustering run.}

\item{initMethod}{A character vector indicating the initialization method
to be used. Default is "kmeans". Options are currently "kmeans" only.}

\item{modelNames}{A character string indicating the model name to generate
covariance matrix, Sigma. Since the largest contribution of free parameters
is through the covariance matrices, it is the focus for introduction
of parsimony. The constraints can be imposed on Lambda (loading matrix)
and Psi (error variance and isotropic) which are used to generate Sigma.
The 'C' stands for constrained and 'U' stands for unconstrained. The order
goes as loading matrix (Lambda), error variance (Psi) and isotropic (Psi),
respectively. Example, if the loading matrix (Lambda), error variance (Psi)
and isotropic are all constrained, then select 'CCC'. Options are "CCC",
"UUU", or "UCC".}

\item{normalize}{A string with options "Yes" or "No" specifying
if normalization should be performed. Currently, normalization factors
are calculated using TMM method of edgeR package. Default is "Yes".}
}
\value{
Returns an S3 object of class mixMPLNFA with results.
\itemize{
\item dataset - The input dataset on which clustering is performed.
\item dimensionality - Dimensionality of the input dataset.
\item normalizationFactors - A vector of normalization factors used
for input dataset.
\item gmin - Minimum number of components/clusters considered in the clustering
run.
\item gmax - Maximum number of components/clusters considered in the clustering
run.
\item pmin - Minimum number of latent factors (p) considered in the clustering
run.
\item pmax - Maximum number of latent factors (p) considered in the clustering
run.
\item allResults - A list with all results. To access results use the format
of 'g, p, model', respectively. See examples.
\item logLikelihood - A vector with value of final log-likelihoods for
each component/cluster size.
\item numbParameters - A vector with number of parameters for each
component/cluster size.
\item trueLabels - The vector of true labels, if provided by user.
\item ICLresults - A list with all ICL model selection results.
\item BICresults - A list with all BIC model selection results.
\item AICresults - A list with all AIC model selection results.
\item AIC3results - A list with all AIC3 model selection results.
\item totalTime - Total time used for clustering and model selection.
}
}
\description{
Performs simultaneous clustering and factor analysis using
parsimonious mixtures of multivariate Poisson-log
normal factor analyzers family (PMPLNFA) via
variational Gaussian approximations. Model selection can
be done using AIC, AIC3, BIC and ICL. In the PMPLNFA framework
restrictions are introduced to the model parameters with the
aim of obtaining parsimonious models, which are sufficiently
flexible for clustering purposes. Since the largest contribution
of free parameters is through the covariance matrices, it is
the focus for introduction of parsimony. The model simultaneously
performs factor analysis and cluster analysis, by assuming that
the discrete observed data have been generated by a factor
analyzer model with continuous latent variables.
}
\details{
Starting values (argument: initMethod) play an
important role in the successful operation of this algorithm.
}
\examples{
\dontrun{
  modelChoice <- c("UUU", "UUC", "UCU", "UCC", "CUU", "CUC", "CCU", "CCC")
}

# Example 1
trueMu1 <- c(6.5, 6, 6, 6, 6, 6, 6, 6, 6)
trueMu2 <- c(2, 2.5, 2, 2, 2, 2, 2, 2, 2)

trueSigma1 <- diag(length(trueMu1)) * 2
trueSigma2 <- diag(length(trueMu1))

# Generating simulated data with 2 clusters
sampleData <- MPLNClust::mplnDataGenerator(nObservations = 2000,
                     dimensionality = length(trueMu1),
                     mixingProportions = c(0.79, 0.21),
                     mu = rbind(trueMu1, trueMu2),
                     sigma = rbind(trueSigma1, trueSigma2),
                     produceImage = "No")

 dim(sampleData$dataset) # a dataset of size 2000 by 9

# Clustering
MPLNFAResults <- mixMPLNFA::MPLNFAClust(
                     dataset = sampleData$dataset,
                     membership = sampleData$trueMembership,
                     gmin = 1,
                     gmax = 3,
                     pmin = 1,
                     pmax = 2,
                     modelNames = c("CCU", "UUU"),
                     normalize = "Yes")

names(MPLNFAResults) # see all names of outputs

# To see BIC results
MPLNFAResults$BICresults
MPLNFAResults$BICresults$BICmodelselected

# Compare with true labels
table(MPLNFAResults$BICresults$BICmodelSelectedLabels,
      sampleData$trueMembership)

# Access all results for g = 2, p = 1, model = "UUU"
# UUU is mentioned in second placed for input string c("CCU", "UUU")
MPLNFAResults$allResults[[2]][[1]][[2]]


}
\references{
Aitchison, J. and C. H. Ho (1989). The multivariate Poisson-log normal distribution.
\emph{Biometrika} 76.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.

Arlot, S., Brault, V., Baudry, J., Maugis, C., and Michel, B. (2016).
capushe: CAlibrating Penalities Using Slope HEuristics. R package version 1.1.1.

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions
on Pattern Analysis and Machine Intelligence} 22.

Bozdogan, H. (1994). Mixture-model cluster analysis using model selection criteria
and a new informational measure of complexity. In \emph{Proceedings of the First US/Japan
Conference on the Frontiers of Statistical Modeling: An Informational Approach:
Volume 2 Multivariate Statistical Modeling}, pp. 69–113. Dordrecht: Springer Netherlands.

Robinson, M.D., and Oshlack, A. (2010). A scaling normalization method for differential
expression analysis of RNA-seq data. \emph{Genome Biology} 11, R25.

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Silva, A. et al. (2019). A multivariate Poisson-log normal mixture model
for clustering transcriptome sequencing data. \emph{BMC Bioinformatics} 20.
\href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-2916-0}{Link}
}
\author{
{Anjali Silva, \email{anjali@alumni.uoguelph.ca}, Andrea Payne, \email{andreapayne@cmail.carleton.ca},
Sanjeena Dang, \email{sanjeenadang@cunet.carleton.ca}. }
}
