% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplnFAVisualize.R
\name{mplnVisualizeLine}
\alias{mplnVisualizeLine}
\title{Visualize Clustered Results Via Line Plots}
\usage{
mplnVisualizeLine(
  dataset,
  clusterMembershipVector = NA,
  fileName = paste0("Plot_", date()),
  LinePlotColours = "black",
  printPlot = TRUE,
  format = "pdf"
)
}
\arguments{
\item{dataset}{A dataset of class matrix and type integer such that
rows correspond to observations and columns correspond to variables.}

\item{clusterMembershipVector}{A numeric vector of length nrow(dataset)
containing the cluster membership of each observation. If not provided,
all observations will be treated as belonging to one cluster. Default is NA.}

\item{fileName}{Unique character string indicating the name for the plot
being generated. Default is Plot_date, where date is obtained from
date().}

\item{LinePlotColours}{Character string indicating if the line plots
should be multicoloured or monotone, in black. Options are
'multicolour' or 'black'. Default is 'black'.}

\item{printPlot}{Logical indicating if plot(s) should be saved in local
directory. Default TRUE. Options TRUE or FALSE.}

\item{format}{Character string indicating the format of the image to
be produced. Default 'pdf'. Options 'pdf' or 'png'.}
}
\value{
Plotting function provides the possibility for line plots.
}
\description{
A function to visualize clustering results via line plots.
Each cluster will have its own plot. Data is log-transformed
prior to visualizing. Values for each sample are connected
by dashed lines to illustrate the trends (log counts). The
yellow line shows the mean value (log counts) for each cluster.
}
\examples{
# Example 1
trueMu1 <- c(6.5, 6, 6, 6, 6, 6, 6, 6, 6)
trueMu2 <- c(2, 2.5, 2, 2, 2, 2, 2, 2, 2)
trueMu3 <- c(1, 1, 1, 1, 1, 1, 1, 1, 1)

trueSigma1 <- diag(length(trueMu1)) * 2
trueSigma2 <- diag(length(trueMu1)) * 0.05
trueSigma3 <- diag(length(trueMu1)) * 0.01

# Generating simulated data with 3 clusters
sampleData <- MPLNClust::mplnDataGenerator(nObservations = 2000,
                     dimensionality = length(trueMu1),
                     mixingProportions = c(0.2, 0.3, 0.5),
                     mu = rbind(trueMu1, trueMu2, trueMu3),
                     sigma = rbind(trueSigma1, trueSigma2, trueSigma3),
                     produceImage = "No")

 dim(sampleData$dataset) # a dataset of size 2000 by 9

# Clustering
MPLNFAResults <- mixMPLNFA::MPLNFA(
                     dataset = sampleData$dataset,
                     membership = sampleData$trueMembership,
                     gmin = 1,
                     gmax = 4,
                     pmin = 1,
                     pmax = 1,
                     modelNames = c("CCU", "UUU"),
                     normalize = "Yes")

 # Visualize data using line plot
 # Use navigation buttons to see previous plots
 MPLNFABlack <- mixMPLNFA::mplnVisualizeLine(dataset = sampleData$dataset,
                                         clusterMembershipVector =
                                         MPLNFAResults$BICresults$BICmodelSelectedLabels,
                                         fileName = 'Example1',
                                         printPlot = FALSE)


 # Visualize data using line plot with multicolours
 # Use navigation buttons to see previous plots
 MPLNLineColor <- mixMPLNFA::mplnVisualizeLine(dataset = sampleData$dataset,
                                         clusterMembershipVector =
                                         MPLNFAResults$BICresults$BICmodelSelectedLabels,
                                         fileName = 'Example1MultiColor',
                                         LinePlotColours = "multicolour",
                                         printPlot = FALSE)

 # Example 2
trueMu1 <- c(6.5, 6, 6, 6, 6, 6, 6, 6, 6)
trueMu2 <- c(1, 2.5, 2, 2, 2, 2, 1, 1, 1)

trueSigma1 <- diag(length(trueMu1)) * 2.5
trueSigma2 <- diag(length(trueMu1)) * 0.5

# Generating simulated data with 2 clusters
sampleData2 <- MPLNClust::mplnDataGenerator(nObservations = 2000,
                     dimensionality = length(trueMu1),
                     mixingProportions = c(0.6, 0.4),
                     mu = rbind(trueMu1, trueMu2),
                     sigma = rbind(trueSigma1, trueSigma2),
                     produceImage = "No")

 dim(sampleData2$dataset) # a dataset of size 2000 by 9

# Clustering
MPLNFAResults2 <- mixMPLNFA::MPLNFA(
                     dataset = sampleData2$dataset,
                     membership = sampleData2$trueMembership,
                     gmin = 1,
                     gmax = 3,
                     pmin = 1,
                     pmax = 2,
                     modelNames = c("CCU"),
                     normalize = "Yes")

 # Visualize data using line plot with multicolours
 # Use navigation buttons to see previous plots
 MPLNLineColor <- mixMPLNFA::mplnVisualizeLine(dataset = sampleData2$dataset,
                                         clusterMembershipVector =
                                         MPLNFAResults2$BICresults$BICmodelSelectedLabels,
                                         fileName = 'TwoClusterModel',
                                         LinePlotColours = "multicolour",
                                         printPlot = FALSE)


}
\author{
Anjali Silva, \email{anjali@alumni.uoguelph.ca}
}
